packages/enterprise-foo-construct/pyproject.toml
[project]
name = "enterprise-foo-construct"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.244.0,<3.0.0",
    "enterprise-constructs-base",
]
